[#2] Refactor: NetworkManager 분리
